Driver for smt160 working with interrupts but incorrect values (50% duty all the time).
using 1 channel, switching rising and falling edge, but probably should disable channel before such switch - see manual.

+++ support/kinetis/smt160_kl25.h
@@ -15,7 +15,7 @@
 #define SMT160_KL25_H_
 
 void smt160_init();
-int	 smt160_get_temp();		
+uint16_t smt160_get_temp();		
 
 
 #endif /* SMT160_KL25_H_ */
